sys: Initial implementation of threads and RR scheduler
The initial implementation of threads and scheduler.
The scheduler is a round-robin implementation based on ready
threads pool that is single linked list.
Current thread is stored in g_current_thread. That means for
RR scheduler next thread is always head of the ready pool.

More documentation about the design and implementation will be
provided.

Signed-off-by: Piotr Pryga <pryga.piotr@outlook.com>